Fix `class CanvasPencilDrawingArraysTests`
